+++ conduit/main.py
@@ -2,19 +2,19 @@ import os
 import sys
 
 try:
-    from purplepy.config import STRFRY_PATH
+    from conduit.config import STRFRY_PATH
 except ModuleNotFoundError:
-    print("Run with command: python -m purplepy.main")
+    print("Run with command: python -m conduit.main")
     sys.exit()
 
-from purplepy.db import (
+from conduit.db import (
     client,
     create_weaviate_event_class,
     create_weaviate_user_class,
     load_events_into_weaviate,
     load_neo4j_data,
 )
-from purplepy.parse import parse_event_json
+from conduit.parse import parse_event_json
 
 print("Errors logged in 'logfile.log'")
 # These commented commands need to be run before executing the python code
@@ -38,5 +38,5 @@ parse_event_json(os.path.join(STRFRY_PATH, "dbdump.jsonl"))
 # loads follows, mentions, reactions, replys, users into neo4j
 load_neo4j_data()
 
-# cd [path/to/purple-py]
-print("Start app with 'python -m purplepy.app'")
+# cd [path/to/conduit]
+print("Start app with 'python -m conduit.app'")
