+++ conduit/vectorize.py
